requirement.txt
astroid==2.4.1
autopep8==1.5.2
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.5.1
PyYAML==5.3.1
selenium==3.141.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
